requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn>=0.34.0,<0.35.0
python-dotenv>=1.0.0,<1.1.0
pydantic>=2.0.0,<3.0.0
sentence-transformers>=3.0.0
nltk>=3.8.0
numpy>=1.22.0
torch>=2.0.0 